bot/requirements.txt
python-telegram-bot>=21.0
notion-client>=2.2
python-dotenv>=1.0.0
